requirements.txt
# Personal AI Employee System - Python Dependencies
# Install with: pip install -r requirements.txt

# Core
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0.0

# Web (for Railway deployment)
fastapi>=0.104.0
uvicorn>=0.24.0

# Database
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Gmail API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0

# File watching
watchdog>=3.0.0

# HTTP client
httpx>=0.25.0
aiohttp>=3.9.0

# WhatsApp via Twilio (Gold tier)
twilio>=9.0.0

# Browser automation (optional)
# playwright>=1.40.0

# MCP Servers
mcp>=1.0.0

# Scheduling
schedule>=1.2.0
apscheduler>=3.10.0

# CLI and formatting
rich>=13.0.0
typer>=0.9.0

# Logging
structlog>=23.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0
